requirements.txt
py-cord==2.5.0
pyTelegramBotAPI==4.19.1
python-dotenv==1.0.1
requests==2.32.3
